requirements.txt
# ─── Core Framework ───────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ─── Database ─────────────────────────────────────────────────────
sqlalchemy==2.0.36
# For PostgreSQL support (optional, uncomment if needed):
# psycopg2-binary==2.9.10

# ─── Authentication ───────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# ─── Validation ───────────────────────────────────────────────────
pydantic==2.10.1
pydantic-settings==2.6.1
email-validator==2.2.0

# ─── HTTP Client (for optional OpenAI integration) ────────────────
httpx==0.28.0

# ─── Testing ──────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.0  # also used as async test client

# ─── Development ──────────────────────────────────────────────────
python-dotenv==1.0.1
